move from production code the PathValidator to Tests folder
